Fix subtle bug in createTagging function
We were assuming that the last pushed tag would always been the first to
succeed. This is not necessarily true, and is probably quite unlikely
(the most probable is that tags in the queue behave in a FIFO manner).

So, instead of just popping, find the exact element we want to remove
and remove it.

Signed-off-by: Wincent Colaiuta <win@wincent.com>