Move bower_components into vendor/assets/
Keep clutter out of the top level, as suggested here:

  https://shellycloud.com/blog/2013/09/how-to-manage-front-end-packages-in-rails-with-bower

Somehow, this breaks font serving in development (not sure about
production). Specifically, we're requesting the font files from /fonts/,
but it's 404'ing (in the past, the generated URL was to
/dev-assets/$FONT.$EXT so I am not sure why it's changed).

In prod, we used to generate URLS like
https://$CDN_HOST/assest/$FONT-$DIGEST.$EXT, so it's unclear whether
this will do the right thing in prod.

Reverting my patches to the FontAwesome source, we get out
/dev-assets/$FONT.$EXT URLs again, but requests for that URL die with a
"route not recognized" error.

Oh, asset pipeline...

Signed-off-by: Wincent Colaiuta <win@wincent.com>