Set up base vertical rhythm functionality
I wasn't particularly sold on the vertical rhythm mix-in at Causes, as
it added a layer of abstraction and indirection with little tangible
benefit that I could see. With the latest Compass, however, the ability
to use rem units with a robust px fallback makes this a nice set of
mix-ins indeed.

Here we replace the existing baseline set-up with a call to
establish-baseline. There should be no visual difference at this point,
given the Compass defaults (base-font-size of 16px and base-line-height
of 24px). If I were doing this manually I would have used 100% and 1.5
respectively for those, but convention over configuration wins out.

Signed-off-by: Wincent Colaiuta <win@wincent.com>