Add modular-scale Sass extension
This will replace my hand-coded modular scale stuff with something a
little more community-supported:

  https://github.com/Team-Sass/modular-scale

I'm going with a Bower install rather than a Gem in order to minimize
the involvement of Compass (which at times isn't stable or up-to-date)
and to keep the Gemfile as slim as possible. Hopefully this is lowest
overhead solution.

Signed-off-by: Wincent Colaiuta <win@wincent.com>