Overhaul modular scaling on headings
This is much nicer, and will be nicer still once I get the vertical
rhythm stuff in place more pervasively. As @lencioni says, these
abstractions are all about enabling us to move faster by systematizing a
set of decisions that we would otherwise have to deal with on a
case-by-case basis.

Signed-off-by: Wincent Colaiuta <win@wincent.com>