Fix mangling caused by syntax-highlighting edge case
See:

  https://wincent.com/issues/2058

The root cause here was the non-guaranteed evaluation order of object
properties in JS, which meant that our "string-literal" rule was free to
royally screw things up.

Commenting out the offending rule makes the problem go away, but it is
not a robust fix.

This commit instead swaps out the objects for arrays, so that we can do
an ordered iteration over the rules.

I took the opportunity to move the "safety" rules (for skipping links
and entities) up to the top into a "default" rule block that all
the rulesets inherit from.

I'm in a bit of a hurry, so I am not 100% sure that this is
future-proof, but cursory testing indicates that it works.

Signed-off-by: Wincent Colaiuta <win@wincent.com>