Simplify a regex
We don't have to worry about hyperlinks in single-quoted strings because
the wikitext translator will always use double-quotes for attribute
values.

For example:

  'foo <a href="http://example.com">http://example.com</a> bar'

We can safely wrap that whole span in a string-literal marker without
worrying about cutting it in half.

Compare that with the dangerous case:

  "foo <a href="http://example.com">http://example.com</a> bar"

We can't just scan to the first " after "foo", as that would end up
mangling out anchor tag.

Signed-off-by: Wincent Colaiuta <win@wincent.com>