Do partial state updates in TagWidget
Face-palm... how could I not have known this was possible until now?

I realized from reading this thread that partial updates of the state
object are possible; React will merge them in:

  https://groups.google.com/d/msg/reactjs/n4gvaBDablg/SPxile1ozaAJ

This makes some of the code a lot simpler, and gets rid of that awkward
pattern of mutating `this.state` only to immediately let React know that
something happened with `this.setState(this.state)` (an ugly pattern
which was repeated multiple times).

Note that this means I had to create a couple of temporary variables, in
the places where I was using `splice()` to mutate arrays, seeing as
splice returns the deleted elements and not the array itself. This is
fine, I think, as it means that the data in the state object (that is,
the content of its properties) is truly immutable, which might be
leverageable in the future for some performance benefits.

So, there's the first big pay off from reading the reactjs Google group.

Signed-off-by: Wincent Colaiuta <win@wincent.com>