Revamp issues form styling
This is the first redesigned form (note: all others are horribly broken
with the removal of the default form styles from _display.scss).

This is inspired by the Causes form design. I am not convinced that it's
the best design (takes up a lot of vertical space), but I wanted to
commit it anyway as a snapshot of what it might look like.

Should improve/implement:

  - focus styling for text inputs and textareas
  - think about reducing vertical space by grouping narrow fields onto
    rows (eg. two selects next to one another, on wide screens)

Signed-off-by: Wincent Colaiuta <win@wincent.com>