Add focused form styling, with example
On focusing/blurring an input, we find the corresponding .field-row div
(using JS) and add/remove a .focused class. This what we'll have to do
until browsers add support for "level 4" selectors, such as ".field-row!
> input:focused" (which would allow us to target the .field-row that is
> the parent of a focused input).

The focus consists of a slightly darker background as well as an
emphasized drop shadow (on all edges of the row, not just the top).

I was thinking about not showing a focus on clicking field rows with
checkboxes in them, but with this styling I think it actually makes
sense as valuable feedback (otherwise the user may not notice they've
toggled the checkbox).

As a bonus, includes a tweak to row styling (slightly rounded corners)
and an example form (for posting comments).

Signed-off-by: Wincent Colaiuta <win@wincent.com>