Fix issue moderation
Even since adding another response format to the controller, we haven't
been able to moderate issues as ham. This, despite the fact that the
underlying Ajax call has always specified `dataType: 'json'`.

Let's just cut our losses and let the implementations share code, for
now. It has always been a bit messy; would be nice to clean it up in the
future (perhaps with a faux-resource like ModelModerationsController or
IssueModerationsController or something).

Fixes:

  https://wincent.com/issues/2139

Signed-off-by: Wincent Colaiuta <win@wincent.com>