Revamp user forms
As you can see, this involved some hacking to override unwanted WebKit
interference.

Also note the non-orthodox "nested" form here (not really a form, just
links). Rather than clutter up the CustomFormBuilder with a method to
handle this kind of thing (could be a riff on fieldset, or a "row"
method etc), let's just hand-code it for now until a useful abstraction
becomes evident.

Signed-off-by: Wincent Colaiuta <win@wincent.com>