Update login form
In order to benefit from the CustomFormBuilder goodness, I had to switch
from `form_tag` to `form_for` (they alternative would be to replicate
the markup structure created by the builder, which would be brittle).

So, I created a new `Session` model that is not database-backed and
exists solely for the purpose of modelling this login interaction. Later
on we might be able to do some more refactoring to get some of the logic
out of the controller and into that model (although access to the
underlying session and cookie objects is required, so that puts a pretty
hard limit on what can be conveniently moved).

Signed-off-by: Wincent Colaiuta <win@wincent.com>