Make it easier to use factories from the command line
Just:

  require './spec/support/factory_girl'

And then `make{,!}` at will.

If you fire up a Rails console normally (ie. with `rails c`), this will
happen automatically, by virtue of the `.irbrc` file at the repo root,
which this commit also updates (it's been silently failing for a while).

Note that in the context of RSpec, the factory definitions get found
automatically, thanks to the Railtie that comes with factory_girl_rails.
In the console, however, we have to trigger their loading explicitly;
that's what the conditional at the bottom of the factory_girl.rb file is
for.

Signed-off-by: Wincent Colaiuta <win@wincent.com>