Robustify field_error_proc
I could have gone with something simple like:

  http://stackoverflow.com/a/8380400

But that is just a little too sloppy. For example, I could break it with
content like:

  <textarea>
  class="yes!
  </textarea>

But still, those ugly regexen deserve to be ripped out. They are hard to
read. So, let's justify Nokogiri's ongoing presence in the Gemfile (it
has been dead since we removed that Twitter code a few days ago), let's
use it to robustly parse the HTML tag and rebuild it with the desired
class attributes.

Signed-off-by: Wincent Colaiuta <win@wincent.com>