Remove Ruby 2.1.0 bug workaround
I've added `ulimit -n 2048` to my ~/.zshlocal on this machine, so this
workaround is no longer needed.

The default limit on OS X is very low, about 256, so this makes it
something more reasonable. On my remote Linux boxes, the limit is
1024.

I'm not too worried about this any more, as GC should be happening often
enough in production that there's no need to worry about the limit.

Signed-off-by: Wincent Colaiuta <win@wincent.com>