Allow pill selection by clicking
This is a great day for the republic, as it brings us to feature parity
with the Google autocomplete widget, which I consider to be the nicest
one out there.

In this commit we add pill selection by clicking. The widget gets a new
"handleTagSelect" handler for this purpose. To avoid clashing with the
pre-existing handler of the same name, I gave it a more descriptive name
("handleAutocompleteSelect").

In order for the various click handlers to play nicely with one another,
had to make these tweaks:

  - the pill not only prevents the click default when clicking on the
    close widget, it also stops propagation; this prevents things from
    getting confused in the widget
  - in the pill's click handler, we only need to stop propagation, not
    also prevent the default action (because there is no default action
    on a span).

Signed-off-by: Wincent Colaiuta <win@wincent.com>