On tabbing into TagWidget, forward focus to TagInput
This is pretty subtle and nasty. My first attempt at this was to do
something like this in the TagWidget handleFocus() method:

  if (typeof this.state.selectedPillIndex === "undefined") {
    // we were probably tabbed into from outside the widget...
  }

It turns out that this breaks the "press left to select tag pill"
behavior. In handleLeftDown(), there is code like this:

  this.setState({ selectedPillIndex: this.state.tags.length - 1 });
  this.getDOMNode().focus();

It turns out that we can't rely on the state that is set here to be
visible in the handleFocus() method. My guess is that React's batching
prevents it from becoming immediately visible. This is what happens
(verified in the console):

  console.log(this.state.selectedPillIndex); // prints: undefined
  this.setState({ selectedPillIndex: 3 });
  console.log(this.state.selectedPillIndex); // still prints: undefined

So, the hack here is to store the flag directly on the TagWidget object
itself. It feels "out of band", but I can't see any other way to do
this.

Hopefully the comment will survive long enough that a `git blame` will
lead people here to this explanation and it will make sense. (I'm
reluctant to include this much verbosity in a code comment.)

Signed-off-by: Wincent Colaiuta <win@wincent.com>