Get shift-tab working in tag widget
The trick here is to catch the event on the inner text input, and let
the containing widget know about it via a callback. The widget forces
focus onto the widget element itself and synthesizes another keyboard
event that causes the focus to really jump to the previous element.

Counter-intuitively, this only works when we don't
preventDefault/stopPropagation on the original event.

Signed-off-by: Wincent Colaiuta <win@wincent.com>