Make naming conventions more consistent
Use names like "handleClick", "handleKeydown" for functions that handle
native events.

Use names like "onFoo", "onBar" for functions that serve as callbacks
for child-to-parent communication.

The only wrinkle that I am not sure how to iron out is in the TagPill
class, where I have two elements that specify onClick handlers; one of
them "wins" and gets to use "handleClick" as the name for its handler
(the outermost element), and the other one gets a more "semantic" name:
"handleDelete" (ie. because it deletes pills).

I could call this thing "handleCloseWidgetClick" or something more
descriptive (which would probably be pretty ugly), or I could turn the
inner element into another React class and move the concern out of this
file (but that would harm locality; we're really just talking about two
lines of code here). So, I'll have to meditate on it.

Signed-off-by: Wincent Colaiuta <win@wincent.com>