Set up Browserify
This is how we're going to get React working with Rails. Not using the
react-rails gem as it wants to pull in execjs, which is a pain in the
ass for deployment.

Also, RubyGems and the Ruby ecosystem in general are ridiculously
volatile and unstable. Trade the "devil we know" for the "devil we
don't" and use NPM packages instead.

As far as the asset pipeline is concerned, just dump a big bundle in its
lap and limit its involvement to concatenating that and minifying it for
us.

Note, this means I'm checking in generated code (the bundle.js) and also
bloating the repo a bit with a bunch of locally installed node modules.
Hopefully this is ok.

Installed with:

  npm install --save-dev browserify watchify reactify

Note, couldn't do:

  npm install --save react

For some reason, this kept downloading the react-page package... So I
manually edited the package.json file instead.

To build:

  npm run bundle

To build continuously

  npm start

May also want to check out some of the links here:

  http://lincolnloop.com/blog/untangle-your-javascript-browserify/

in the future. Would be nice to get development mode source maps
working, although we can probably live without them for now.

Note: as a bonus of using reactify, which in turn uses the react-tools
package, which in turn bundle Facebook's jstransform, we can start using
ES6 features in the codebase right now.

The original wonderful idea to do this without building on the teetering
stack of cards that is Rails and the gem ecosystem here:

  http://learnjs.io/blog/2014/03/17/using-browserify-with-rails/

Signed-off-by: Wincent Colaiuta <win@wincent.com>