Migrate some Bower components to Browserify (node_modules)
There is probably still a use for Bower for managing non-JS
resources, but Browserify is the new hotness for us for the JS stuff, as
it will eventually allow us to move towards the holy grail of using the
same rendering code on the server and the client.

So, jQuery, Underscore and spin.js migrate over. Backbone dies, as I
wasn't using it (was only going to use it). React also goes away, as
there is no need for it to be in two places at once.

My class.js implementation goes away, as I can use "real" ES6 classes
(thanks to FB's jstransform). The Spinner and Menu classes get migrated
to classes.

Signed-off-by: Wincent Colaiuta <win@wincent.com>