Update assets/deploy processes to automate Browserify bundling
It's likely a smell that our "check" actually updates the bundle, but
the nature of the beast is that's the only way we _can_ check if the
bundle is up-to-date.

I guess I could refrain from clobbering the in-tree copy of the bundle,
but this is the best my sleep-deprived mind can come up with for now.
The other main alternative is to make the rake task just plain
`assets:browserify` instead of `assets:browserify:check`, which would be
a bit less misleading.

Signed-off-by: Wincent Colaiuta <win@wincent.com>