Attempted workaround for annoying Sprockets/Sass bug
The comment in the Gemfile says it all. Rather than going to the hassle
of making a submodule, this is just an inline clone:

  mkdir vendor/gems
  cd !$
  git clone https://github.com/sstephenson/sprockets.git
  cd sprockets
  git co v2.11.0
  git cherry-pick 655f129fa910f7d46803fdc66d6d5de53a85f210
  rm -rf .git
  cd ../../..
  rm -r tmp/cache # just in case...

That supposedly silences the spewing console messages like:

  Warning. Error encountered while saving cache 7471105c2a0c60ce122655b13b4296aa53117392/_TagWidget.scssc: can't dump anonymous class #<Class:0x007fa95b1a5750>

But it also means we aren't caching now so things will likely be slower.

:-(

But it gets worse... The console is still getting spammed, _and_ updates
to Sass files aren't being picked up, making it impossible to do any
development which involves Sass. Fucking great.

What a ridiculous, fucked-up ecosystem. I can't get off of this soon
enough.

Signed-off-by: Wincent Colaiuta <win@wincent.com>