Partial stab at progressive enhancement
This doesn't quite work, but almost does.

When JS is active, we wind up with this in the DOM:

  <input type="text" value="" name="articles[pending_tags]" />
  <input type="hidden" value="foo bar baz" name="articles[pending_tags]" />

Form submission works, the record gets created with tags.

If I turn JS off after the initial render, I wind up with this in the
DOM:

  <input type="text" value="foo bar baz" name="articles[pending_tags]" />
  <input type="hidden" value="" name="articles[pending_tags]" />

This doesn't work. The latter value ("") trumps the former.

Will need to think on this a little more.

Signed-off-by: Wincent Colaiuta <win@wincent.com>