Progressively enhance TagWidget
They said it couldn't be done.

See the inline comments for more details.

There is a lot of ugly hoop-jumping involved here, and I am not sure if
it's worth it. Nevertheless, committing this an example of how it can be
done. I hope that other components aren't quite so fiddly as this.

There is some redundant re-rendering involved thanks to those `setState`
calls, and we need a `noJS` class (helps us with specificity) in order
to set a small width on the text input; without this, when JS is turned
on, we somehow wind up with the text input at 100% width (ie. too wide)
until the first input event.

Signed-off-by: Wincent Colaiuta <win@wincent.com>