Add grunt and node-jsx
This is to make it easier to develop server-side rendering code. With
it, you can try out React components in the `node` REPL:

  require('node-jsx').install({harmony: true})
  React = require('React')
  TagPill = require('./app/assets/javascripts/TagPill')
  React.renderComponentToString(TagPill())

I also did `npm install -g react-tools`, to give me global access to the
`jsx` CLI tool. With this, I can do:

  jsx --harmony app/assets/javascripts server/build

And all the modules I need on the server will be processed for me and
wind up in `server/build`. Note that `bundle.js` also winds up in there,
which isn't what I want, so I might have to exclude this in some way.

Note that for either of these to work, I needed `@jsx` annotations in
all of my files which use JSX.

Next steps will be getting this working via Grunt, which will make this
a little bit easier to deal with and will eventually give me something
to plug into the deploy process.

Signed-off-by: Wincent Colaiuta <win@wincent.com>