Try to unbreak assets pipeline
Sigh... Somehow, the assets pipeline has decided to ignore all of my
Sprockets `//= require` statements.

I decided to inline wincent.js and products.js to at least remove the
corresponding requires as points of failure, and that "fixed" that
problem, but now the bundle is not being required either.

No idea what's going on, but it's annoying.

Signed-off-by: Wincent Colaiuta <win@wincent.com>