Switch to Ruby 2.1.2
Ok, let's try this again. We'll need to anyway to target what's
currently on the new AWS box. If I can get this building, let's try
2.0.0-p451 again...

Or maybe not. Ruby is just clowny enough that I'd hate not to be able to
update to the latest and greatest due to a critical security
vulnerability that Amazon is slow to patch... Will need to think about
it some more.