Implement form changes
Spam honey pot. Adds a (hidden) "website_address" form to all forms and
if a bot fills it in we silently "accept" the submission.

There were two ways to add the form input; this:

  ::ActionView::Helpers::FormTagHelper.class_eval do
    def extra_tags_for_form_with_additions(html_options)
      extra_tags_for_form_without_additions(html_options) +
        hidden_field_tag('thing', 'bar')
    end
    alias_method_chain :extra_tags_for_form, :additions
  end

Or:

  # append a spam-catching field: http://stackoverflow.com/a/2112364
  form_for_without_custom(record, options) do |f|
    concat(hidden_field_tag('extra', 'info'))
    proc.call(f)
  end

Went with the latter as we already had an alias method chain in place
that I could use.