Don't double-escape ampersands from paginator
Not sure whether this will break down under some circumstances, but it
was definitely broken before because of double-escaping turning:

  foo=1&amp;bar=1

into:

  foo=1&amp;amp;bar=1

If I search for the text "&", it winds up percent encoded in the URL as:

  issue%5Bsummary%5D=%26

So I'm guessing something changed somewhere along the long chain of
Rails upgrades. I'd explicitly changed this in f70de8873d (2010), and it
seems like I tested it back then, but it stopped working.