Believe in assets:browserify:check again
I had this commented out for a while because it was raising false
positives (separate runs would report a changed bundle, but the changes
were spurious).

I just did a quick test, and at least for my two unscientific samples,
the output appears to be stable again. Hopefully this is a consequence
of switching to Babel, and not just stupid luck.

As I never committed the commenting-out, what you're seeing here is just
some extra hinting that I've added to the error output. Going to try a
test deploy now...