Cure issue search form of ugliness and amnesia
Switch to using our custom form builder (which we can easily do thanks
to the fact that we already had a model sticking around).

This gives us a pretty form, and means that form values get remembered
across page loads when you hit "search again".

Note that I had to do a bit of hackery to make sure the desired behavior
came across:

- no pollution of GET URL with snowman bullshit
- no pollution of GET URL with submit button name
- no pollution of GET URL with my spam honeypot param (required me to
  pass a new option to the custom form builder)

The only place where I couldn't get parity is the unwanted "new_class"
id and class added to the form, which apparently can't be suppressed.