Sans-serif-ize websafe fonts
And for comparison, this is the sans-serif version of my websafe font
challenge.

It is clearly less readable than the serif one, but it looks so much
better. I think the difference is particularly strong for buttons and
nav. If I could style those independently then I might wind up with a
best of both worlds kind of situation.